refactor: modularize backup management
